Creado ScanViewModel
